Upgrade movement

+++ client/includes/renderable_entity.h
@@ -27,8 +27,8 @@ private:
     bool is_moving;
     float movement_speed;
 
-    uint8_t body_id;
-    uint8_t head_id;
+    uint16_t body_id;
+    uint16_t head_id;
     uint8_t weapon_id;
     uint8_t shield_id;
 
@@ -36,7 +36,7 @@ private:
 
 public:
     RenderableEntity(uint32_t id_, EntityType type_, int start_tile_x_, int start_tile_y_,
-                     uint8_t body_id_, uint8_t head_id_, uint8_t weapon_id_, uint8_t shield_id_,
+                     uint16_t body_id_, uint16_t head_id_, uint8_t weapon_id_, uint8_t shield_id_,
                      bool is_short_race_ = false);
 
     virtual ~RenderableEntity() = default;
